// src/Models/Models.ts
// Generated by https://quicktype.io

export interface IForecast {
  pronostico?: IForecast[];
  location: Location | null;
  temp_c: number;
  temp_f: number;
  wind: number;
  wind_dir: null | string;
  condition: string;
  date: string;
  icon: string;
}

export interface Location {
  name: string;
  region: string;
  country: string;
  lat: number;
  lon: number;
  tzId: string;
  localtimeEpoch: number;
  localtime: string;
}

export interface ICurrent {
  location: Location;
  temp_c: number;
  temp_f: number;
  wind: number;
  wind_dir: string;
  condition: string;
  date: string;
  icon: string;
}